creating Hackhlth.md
HackHLTH is the premier hackathon for the health ecosystem, bringing together the most passionate developers, designers, and entrepreneurs with the most forward-thinking sponsors, to solve the industry’s toughest problems. With health at a critical point, the need for new products and services to move forward, faster and more effectively, is more important than ever.

Participants will have 36 hours to collaborate and create meaningful solutions, with the 8 most impactful projects sharing more than $80,000 in prizes.

The industry is looking to you to accelerate innovation and drive meaningful change locally and globally. Register now and join us in Vegas from October 25-27 on our mission to reshape health.